Organizes pages and components into folders
