+++ src/Pages/Home.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import Context from "./Context";
+import Context from "../Context/Context";
 
 export default class Home extends React.Component {
   static contextType = Context;
